Fold `test/pkg.jl` into the topic test files
`pkg.jl` was the pre-rewrite grab bag of operation tests. Move each of
its testsets to the file for its topic (`add.jl`, `develop.jl`,
`update.jl`, `pkgtest.jl`, `status.jl`, `project_files.jl`, `repo.jl`,
`repl.jl`, `concurrent.jl`, `misc.jl`) and put the gc and usage-log tests
in a new `gc.jl`.

Dropped as already covered: the pin/free/update sequence, the `update`
uplevel checks and the wrong/missing UUID checks that only tested for a
throw. The `coverage specific path` testset ran `Pkg.test` with
`--code-coverage=<tracefile>`, which spends about a minute in julia's
LCOV writer just for `using Test`; it now checks the flags that
`gen_subprocess_flags` produces instead.
