+++ test/repo.jl
@@ -128,4 +128,48 @@ end
     end
 end
 
+@testset "protocols" begin
+    isolate() do
+        mktempdir() do devdir
+            withenv("JULIA_PKG_DEVDIR" => devdir) do
+                try
+                    # Test below commented out because it is really slow, https://github.com/JuliaLang/Pkg.jl/issues/1291
+                    #Pkg.setprotocol!(domain = "github.com", protocol = "notarealprotocol")
+                    #@test_throws PkgError Pkg.develop("Example")
+                    Pkg.setprotocol!(domain = "github.com", protocol = "https")
+                    Pkg.develop("Example")
+                    @test isinstalled(TEST_PKG)
+                finally
+                    Pkg.setprotocol!(domain = "github.com")
+                end
+            end
+        end
+        try
+            https_url = "https://github.com/JuliaLang/Example.jl.git"
+            ssh_url = "ssh://git@github.com/JuliaLang/Example.jl.git"
+            @test Pkg.GitTools.normalize_url(https_url) == https_url
+            Pkg.setprotocol!(domain = "github.com", protocol = "ssh")
+            @test Pkg.GitTools.normalize_url(https_url) == ssh_url
+            # TODO: figure out how to test this without
+            #       having to deploy a ssh key on github
+            #Pkg.develop("Example")
+            #@test isinstalled(TEST_PKG)
+
+            https_url = "https://gitlab.example.com/example/Example.jl.git"
+            ssh_url = "ssh://git@gitlab.example.com/example/Example.jl.git"
+
+            @test Pkg.GitTools.normalize_url(https_url) == https_url
+            Pkg.setprotocol!(domain = "gitlab.example.com", protocol = "ssh")
+            @test Pkg.GitTools.normalize_url(https_url) == ssh_url
+
+            @test_deprecated Pkg.setprotocol!("ssh")
+            @test_deprecated Pkg.GitTools.setprotocol!("ssh")
+
+        finally
+            Pkg.setprotocol!(domain = "github.com")
+            Pkg.setprotocol!(domain = "gitlab.example.com")
+        end
+    end
+end
+
 end # module
